backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
python-multipart==0.0.12
pydantic==2.9.2
email-validator==2.1.0
python-dotenv==1.0.1
alembic==1.14.0
geopandas>=0.13.0
rasterio>=1.3.0
shapely>=2.0.0
fiona>=1.8.0
pyproj>=3.4.0
numpy>=1.21.0
pillow>=9.0.0
apscheduler>=3.10.0
aiofiles>=0.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiohttp>=3.8.0